api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arabic-reshaper==3.0.0",
    "asgiref==3.8.1",
    "asn1crypto==1.5.1",
    "brotli==1.1.0",
    "certifi==2025.1.31",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "cpf==2.1",
    "cryptography==44.0.2",
    "cssselect2==0.8.0",
    "defusedxml==0.7.1",
    "django==5.1.7",
    "django-cors-headers==4.7.0",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.5.1",
    "dotenv==0.9.9",
    "faker>=37.11.0",
    "fonttools==4.57.0",
    "gunicorn==23.0.0",
    "html5lib==1.1",
    "idna==3.10",
    "lxml==5.4.0",
    "oscrypto==1.3.0",
    "pathlib==1.0.1",
    "pillow==11.1.0",
    "psycopg2-binary==2.9.10",
    "pycparser==2.22",
    "pydyf==0.11.0",
    "pyhanko==0.26.0",
    "pyhanko-certvalidator==0.26.8",
    "pyjwt==2.10.1",
    "pypdf==5.4.0",
    "pyphen==0.17.2",
    "python-bidi==0.6.6",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",
    "qrcode==8.1",
    "reportlab==4.3.1",
    "requests==2.32.3",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "svglib==1.5.1",
    "tinycss2==1.4.0",
    "tinyhtml5==2.0.0",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "uritools==4.0.3",
    "urllib3==2.4.0",
    "validate-docbr==1.10.0",
    "weasyprint==65.1",
    "webencodings==0.5.1",
    "whitenoise>=6.11.0",
    "xhtml2pdf==0.2.17",
    "zopfli==0.2.3.post1",
]

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["data", "templates"]
